+++ src/general/BirdCarnivorous.java
@@ -4,7 +4,7 @@ public abstract class BirdCarnivorous extends Bird implements Predator {
 
 	private int eatAnimal;
 
-	public BirdCarnivorous(int weight, int lifetime, Sex sex, int growth, int wingspan, int rangeOfFlight,
+	public BirdCarnivorous(double weight, int lifetime, Sex sex, int growth, int wingspan, int rangeOfFlight,
 			boolean canFlight, int eatAnimal, int speed) {
 		super(weight, lifetime, sex, growth, wingspan, rangeOfFlight, canFlight, speed);
 		this.eatAnimal = eatAnimal;
