+++ src/general/BirdGranivorous.java
@@ -4,7 +4,7 @@ public abstract class BirdGranivorous extends Bird {
 
 	private int eatGrown;
 
-	public BirdGranivorous(int weight, int age, Sex sex, int growth, int wingspan, int rangeOfFlight, boolean canFlight,
+	public BirdGranivorous(double weight, int age, Sex sex, int growth, int wingspan, int rangeOfFlight, boolean canFlight,
 			int eatGrown, int speed) {
 		super(weight, age, sex, growth, wingspan, rangeOfFlight, canFlight, speed);
 		this.eatGrown = eatGrown;
