+++ src/general/BirdOmnivorous.java
@@ -4,7 +4,7 @@ public abstract class BirdOmnivorous extends Bird implements JustEat {
 
 	private int eatFood;
 
-	public BirdOmnivorous(int weight, int age, Sex sex, int growth, int wingspan, int rangeOfFlight, boolean canFlight,
+	public BirdOmnivorous(double weight, int age, Sex sex, int growth, int wingspan, int rangeOfFlight, boolean canFlight,
 			int eatFood, int speed) {
 		super(weight, age, sex, growth, wingspan, rangeOfFlight, canFlight, speed);
 		this.eatFood = eatFood;
