+++ src/general/FishCarnivorous.java
@@ -3,7 +3,7 @@ package general;
 public abstract class FishCarnivorous extends Fish implements Predator {
 	private int numberOfTeeth;
 
-	public FishCarnivorous(int weight, int lifetime, Sex sex, int lenght, int speed, Fertilization fertilization,
+	public FishCarnivorous(double weight, int lifetime, Sex sex, int lenght, int speed, Fertilization fertilization,
 			int numberOfTeeth) {
 		super(weight, lifetime, sex, lenght, speed, fertilization);
 		this.numberOfTeeth = numberOfTeeth;
