+++ src/general/FishOmnivorous.java
@@ -3,7 +3,7 @@ package general;
 public abstract class FishOmnivorous extends Fish implements JustEat {
 	private int eatFood;
 
-	public FishOmnivorous(int weight, int age, Sex sex, int lenght, int speed, Fertilization fertilization,
+	public FishOmnivorous(double weight, int age, Sex sex, int lenght, int speed, Fertilization fertilization,
 			int eatFood) {
 		super(weight, age, sex, lenght, speed, fertilization);
 		this.eatFood = eatFood;
