+++ src/general/Insect.java
@@ -3,7 +3,7 @@ package general;
 public abstract class Insect extends Animal {
 	private int quantityLegs;
 
-	public Insect(int weight, int age, Sex sex, int quantityLegs, int speed) {
+	public Insect(double weight, int age, Sex sex, int quantityLegs, int speed) {
 		super(weight, age, sex, speed);
 		this.quantityLegs = quantityLegs;
 	}
