+++ src/general/InsectOmnivorous.java
@@ -3,7 +3,7 @@ package general;
 public abstract class InsectOmnivorous extends Insect implements JustEat {
 	private int range;
 
-	public InsectOmnivorous(int weight, int age, Sex sex, int quantityLegs, int range, int speed) {
+	public InsectOmnivorous(double weight, int age, Sex sex, int quantityLegs, int range, int speed) {
 		super(weight, age, sex, quantityLegs, speed);
 		this.range = range;
 	}
