+++ src/general/MammalHerbivorous.java
@@ -4,7 +4,7 @@ public abstract class MammalHerbivorous extends Mammal implements Run {
 
 	private int quatityOfStomach;
 
-	public MammalHerbivorous(int weight, int age, Sex sex, int height, int lenght, int speed, int quatityOfStomach) {
+	public MammalHerbivorous(double weight, int age, Sex sex, int height, int lenght, int speed, int quatityOfStomach) {
 		super(weight, age, sex, height, lenght, speed);
 		this.quatityOfStomach = quatityOfStomach;
 	}
