+++ src/general/MammalInsectivorous.java
@@ -4,7 +4,7 @@ public abstract class MammalInsectivorous extends Mammal implements JustEat {
 
 	private int mouthparts;
 
-	public MammalInsectivorous(int weight, int age, Sex sex, int height, int lenght, int speed, int mouthparts) {
+	public MammalInsectivorous(double weight, int age, Sex sex, int height, int lenght, int speed, int mouthparts) {
 		super(weight, age, sex, height, lenght, speed);
 		this.mouthparts = mouthparts;
 	}
