+++ src/simple/animal/AntEater.java
@@ -6,7 +6,7 @@ import general.Predator;
 
 public class AntEater extends MammalInsectivorous implements Predator {
 
-	public AntEater(int weight, int age, Sex sex, int height, int lenght, int speed, int mouthparts) {
+	public AntEater(double weight, int age, Sex sex, int height, int lenght, int speed, int mouthparts) {
 		super(weight, age, sex, height, lenght, speed, mouthparts);
 	}
